update api handlers
